@@ -21,7 +21,7 @@ dependencies:
   material_design_icons_flutter: ^5.0.6595
   clipboard: ^0.1.3
   shared_preferences: ^2.0.12
-  font_awesome_flutter: ^9.2.0
+  font_awesome_flutter: ^10.1.0
   sleek_circular_slider: ^2.0.1
   share: ^2.0.4
   sqflite: ^2.0.2
@@ -35,7 +35,7 @@ dependencies:
   auto_size_text: ^3.0.0
   screenshot: ^1.2.3
   get_storage: ^2.0.3
-  audioplayers: ^0.20.1
+  audioplayers: ^1.0.1
   enum_to_string: ^2.0.1
   intl: ^0.17.0
   flutter_zoom_drawer: ^3.0.1
@@ -47,7 +47,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.4
+  flutter_lints: ^2.0.1
 
 flutter_native_splash:
   android: true
