@@ -1,5 +1,5 @@
 import 'package:hisnelmoslem/shared/functions/print.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:url_launcher/url_launcher_string.dart';
 
 sendEmail(
     {required String toMailId,
@@ -7,8 +7,8 @@ sendEmail(
     required String body}) async {
   var url = 'mailto:$toMailId?subject=$subject&body=$body';
   try {
-    if (await canLaunch(url)) {
-      await launch(url);
+    if (await canLaunchUrlString(url)) {
+      await launchUrlString(url);
     } else {
       throw 'Could not launch $url';
     }
