@@ -14,9 +14,9 @@ class ImageVarFontBuilder extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     DashboardController dashboardController = Get.put(DashboardController());
-    String _titleWithIndex =
+    String titleWithIndex =
         "${dashboardController.allTitle[dbContent.titleId - 1].name} | ذكر رقم ${dbContent.orderId}";
-    String _titleWithoutIndex =
+    String titleWithoutIndex =
         dashboardController.allTitle[dbContent.titleId - 1].name;
     return GetBuilder<ShareAsImageController>(builder: (controller) {
       return Card(
@@ -37,8 +37,8 @@ class ImageVarFontBuilder extends StatelessWidget {
                 padding: const EdgeInsets.all(10),
                 child: Text(
                   controller.showZikrIndex
-                      ? _titleWithIndex
-                      : _titleWithoutIndex,
+                      ? titleWithIndex
+                      : titleWithoutIndex,
                   textAlign: TextAlign.center,
                   style: TextStyle(
                     fontWeight: FontWeight.bold,
